requirements.txt
requests>=2.32.0
beautifulsoup4>=4.12.3
lxml>=5.2.2
tenacity>=9.0.0
tqdm>=4.66.4
python-dateutil>=2.9.0.post0